requirements.txt
python-dotenv==1.1.1
aiohttp==3.12.15
livekit==1.0.12
livekit-agents==1.2.5
livekit-api==1.0.5
livekit-plugins-google==1.2.5
livekit-plugins-deepgram==1.2.5
livekit-plugins-elevenlabs==1.2.5
livekit-plugins-silero==1.2.5
livekit-plugins-turn-detector==1.2.5
torch
torchaudio
